+++ lib/src/android/webview_feature.dart
@@ -152,7 +152,8 @@ class AndroidWebViewFeature {
       const AndroidWebViewFeature._internal(
           "SAFE_BROWSING_RESPONSE_SHOW_INTERSTITIAL");
 
-  ///
+  ///Use [SAFE_BROWSING_ALLOWLIST] instead.
+  @Deprecated('Use `SAFE_BROWSING_ALLOWLIST` instead')
   static const SAFE_BROWSING_WHITELIST =
       const AndroidWebViewFeature._internal("SAFE_BROWSING_WHITELIST");